Add new fields to expenses
