+++ src/applications/medical-expense-report/config/chapters/03-additional-information/uploadDocuments.js
@@ -32,7 +32,6 @@ const UploadMessage = (
 export default {
   title: 'Upload documents',
   path: 'expenses/additional-information/upload-documents',
-  depends: formData => formData.hasCareExpenses === true,
   uiSchema: {
     ...titleUI('Submit your supporting documents'),
     'ui:description': Description,
