+++ src/applications/medical-expense-report/tests/fixtures/mocks/local-mock-responses.js
@@ -54,6 +54,8 @@ const responses = {
     return res.json({
       data: {
         attributes: {
+          name: 'test-upload.pdf',
+          size: 15306,
           confirmationCode: '5',
         },
       },
